+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2024-09-09
+
+- Adding option to different page size
+- Renaming `itemsPerPage` prop to `pageSize`
+
 ## [3.1.1] - 2024-08-23
 
 - Fixing ui issue hiding the sort icons